junction-backend/requirements.txt
Flask
pymongo
python-dotenv
openai-whisper
flask_cors
torch
transformers
pika
minio
